pyproject.toml
[tool.poetry]
name = "md-task"
version = "0.1.0"
description = ""
authors = ["noonacedia <ruyearn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.5"
djangorestframework = "^3.15.2"
pytest = "^8.3.4"
coverage = "^7.6.10"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.4"}
ruff = "^0.9.3"
djangorestframework-simplejwt = "^5.4.0"
msgspec = "^0.19.0"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"
pytest-lazy-fixtures = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "W", # pycodestyle
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "COM", # flake8-commas
    "Q", # flake8-quotes
    "C90", # mccabe
    "A", # flake8-builtins
    "FIX", # flake8-fixme
]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
max-complexity = 5

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.core.settings"

[tool.coverage.run]
source = ["./tests"]
omit = [
    "tests/*",
]
